Perception: use enum class rather than enum for object and sensor. (#3024)
* Perception: use enum class rather than enum for object and sensor.

* Navigation: used longer lane from perception lane marker; output empty
navigation path when navigation line size is 0

+++ modules/perception/obstacle/fusion/probabilistic_fusion/pbf_track_object_distance.cc
@@ -39,7 +39,7 @@ float PbfTrackObjectDistance::Compute(
     const PbfTrackPtr &fused_track, const PbfSensorObjectPtr &sensor_object,
     const TrackObjectDistanceOptions &options) {
   const SensorType &sensor_type = sensor_object->sensor_type;
-  ADEBUG << "sensor type: " << sensor_type;
+  ADEBUG << "sensor type: " << static_cast<int>(sensor_type);
   PbfSensorObjectPtr fused_object = fused_track->GetFusedObject();
   if (fused_object == nullptr) {
     ADEBUG << "fused object is nullptr";
